+++ parse_type/count_digits_functions.c
@@ -6,7 +6,7 @@
 /*   By: rnakai <rnakai@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/14 11:53:42 by rnakai            #+#    #+#             */
-/*   Updated: 2020/08/16 13:48:37 by rnakai           ###   ########.fr       */
+/*   Updated: 2020/08/17 11:13:46 by rnakai           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,9 +56,9 @@ int				count_digits_u_int16base(unsigned int num)
 	return (digits);
 }
 
-unsigned long	count_digits_u_long16base(unsigned long num)
+int				count_digits_u_long16base(unsigned long num)
 {
-	unsigned long	digits;
+	int	digits;
 
 	digits = 1;
 	while (num >= 16)
